pyproject.toml
[tool.poetry]
name = "ism_method"
version = "0.1.0"
description = ""
authors = ["Haowei Wang <hw_wang@whu.edu.cn>"]

[tool.poetry.dependencies]
python = "^3.10"
pymongo = "^4.10.1"
tqdm = "^4.66.5"
openai = "^1.52.2"
pyyaml = "^6.0.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
generate = "ism_method.main:main"
write = "ism_method.data_load:main"